refactor(appliance/qdrant_quadlet): modularize Qdrant Quadlet validation into submodules
Reorganized the Qdrant Quadlet validation logic by splitting the original 473-line
file into multiple focused modules: error handling, parsing, tests, and main logic.
This restructuring keeps each source file under the 400-line guideline and improves
code maintainability without changing any validation behavior.
