+++ crates/repovec-core/src/appliance/qdrant_quadlet/parser.rs
@@ -1,0 +1,57 @@
+//! Minimal section-aware parser for the subset of Quadlet syntax used here.
+
+use std::collections::BTreeMap;
+
+use super::QdrantQuadletError;
+
+#[derive(Debug)]
+pub(super) struct ParsedQuadlet {
+    sections: BTreeMap<String, BTreeMap<String, Vec<String>>>,
+}
+
+impl ParsedQuadlet {
+    pub(super) fn parse(contents: &str) -> Result<Self, QdrantQuadletError> {
+        let mut sections = BTreeMap::<String, BTreeMap<String, Vec<String>>>::new();
+        let mut current_section: Option<String> = None;
+
+        for (line_index, raw_line) in contents.lines().enumerate() {
+            let line_number = line_index + 1;
+            let line = raw_line.trim();
+            if line.is_empty() || line.starts_with('#') {
+                continue;
+            }
+
+            if let Some(section) = parse_section_header(line) {
+                current_section = Some(section.to_owned());
+                sections.entry(section.to_owned()).or_default();
+                continue;
+            }
+
+            let Some((key, value)) = line.split_once('=') else {
+                return Err(QdrantQuadletError::InvalidLine { line_number, line: line.to_owned() });
+            };
+
+            let Some(section) = &current_section else {
+                return Err(QdrantQuadletError::PropertyBeforeSection {
+                    line_number,
+                    line: line.to_owned(),
+                });
+            };
+
+            sections
+                .entry(section.clone())
+                .or_default()
+                .entry(key.trim().to_owned())
+                .or_default()
+                .push(value.trim().to_owned());
+        }
+
+        Ok(Self { sections })
+    }
+
+    pub(super) fn values(&self, section: &str, key: &str) -> &[String] {
+        self.sections.get(section).and_then(|entries| entries.get(key)).map_or(&[], Vec::as_slice)
+    }
+}
+
+fn parse_section_header(line: &str) -> Option<&str> { line.strip_prefix('[')?.strip_suffix(']') }
